jest resolution needs to be fixed
