--- packages/backend/src/classes/stateMachine/__tests__/shared.ts
@@ -1,17 +1,0 @@
-import { IClient } from '../../../interfaces';
-
-export const testCurator: IClient = {
-  id: 'curator',
-  sessionId: '',
-  close: () => {
-    return;
-  },
-};
-
-export const testPlayer: IClient = {
-  id: 'player',
-  sessionId: '',
-  close: () => {
-    return;
-  },
-};
